ACCUMULO-19 - provides three packages, and the templates for some more for the future. Accumulo is the bare bones install, accumulo-native will introduce the code neccessary to build the native libraries and build them, and accumulo-test imports all of the testing code under ACCUMULO_HOME/test. There are control files for future files, but they are not developed, but I think they are fairly straightforward. Everything installs to /usr/lib/accumulo, with symlinks for the conf directory pointing to /etc/accumulo/conf (with all example files copied into that directory). Additional symlinks are made from logs and walogs to /var/log/accumulo/log and walog respectively. Remove leaves everything in /var/log/accumulo and /etc/accumulo/ intact. Purge will remove them. Accumulo-native is tied directly to the accumulo version. I have nothing in place with regard to upgrades and things like that, but I don't think that is necessary, at least for now.
The only changes made outside of the pom were to the accumulo and config.sh script, where I introduced readlink -f to the input path to resolve symlinks. This is necessary for the symlink from /usr/bin/accumulo to still work with the scripts.

The installation will also correct the paths for JAVA_HOME, ZOOKEEPER_HOME, and HADOOP_HOME in the -env.sh scripts to the appropriate paths should the processes be installed automatically from bigtop and the debian repositories. The user must still copy the configuration file they want to the default name, but it will be preserved between installations provided there was no purge.



git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/branches/1.4@1293477 13f79535-47bb-0310-9956-ffa450edef68